Move codelists to codelists file
